+++ CHANGES.md
@@ -4,8 +4,9 @@
 
 ---
 
-* @dev (2025-02-08)
+* @dev (????-??-??)
   * Added option to set custom user agent string with `--user-agent` switch.
+  * App name and profile is now included in context menu actions like `Quit <NAME>`
 
 * 1.7.0 (2025-01-23)
   * Tray icon now shows `--name` value as its tooltip when set.