+++ CHANGES.md
@@ -5,6 +5,7 @@
 ---
 
 * @dev (????-??-??)
+  * Added bookmark support with context menu integration.
   * Added option to set custom user agent string with `--user-agent` switch.
   * App name and profile is now included in context menu actions like `Quit <NAME>`